Introduce rpu_config for handling tile parameters
Revamp the way parameters are modelled, centered around "RPU configs".
* add a `aihwkit.simulator.configs` containing the classes that model
  the resistive processing unit configurations.
* Tiles and Layers accept a `FooRPUConfig` instance instead of a
  resistive device.
