+++ .pylintrc
@@ -344,7 +344,7 @@ contextmanager-decorators=contextlib.contextmanager
 # List of members which are set dynamically and missed by pylint inference
 # system, and so shouldn't trigger E1101 when accessed. Python regular
 # expressions are accepted.
-generated-members=get_layer,get_tile,get_resistive_device,bias
+generated-members=get_layer,get_tile,get_rpu_config,bias
 
 # Tells whether missing members accessed in mixin class should be ignored. A
 # mixin class is detected if its name ends with "mixin" (case insensitive).
