+++ src/aihwkit/simulator/configs/configs.py
@@ -1,0 +1,80 @@
+# -*- coding: utf-8 -*-
+
+# (C) Copyright 2020 IBM. All Rights Reserved.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
+
+"""Configurations for resistive processing units."""
+
+from dataclasses import dataclass, field
+from typing import ClassVar, Type
+
+from aihwkit.simulator.configs.devices import (
+    FloatingPointDevice, ConstantStepDevice, PulsedDevice, UnitCellDevice
+)
+from aihwkit.simulator.configs.utils import (
+    BackwardIOParameters, IOParameters, UpdateParameters, tile_parameters_to_bindings
+)
+from aihwkit.simulator.rpu_base import devices
+
+
+@dataclass
+class FloatingPointRPUConfig:
+    """Configuration for a floating point resistive processing unit."""
+
+    device: FloatingPointDevice = field(default_factory=FloatingPointDevice)
+    """Parameters that modify the behavior of the pulsed device."""
+
+
+@dataclass
+class SingleRPUConfig:
+    """Configuration for an analog (pulsed device) resistive processing unit."""
+
+    bindings_class: ClassVar[Type] = devices.AnalogTileParameter
+
+    device: PulsedDevice = field(default_factory=ConstantStepDevice)
+    """Parameters that modify the behavior of the pulsed device."""
+
+    forward: IOParameters = field(default_factory=IOParameters)
+    """Input-output parameter setting for the forward direction."""
+
+    backward: BackwardIOParameters = field(
+        default_factory=BackwardIOParameters)
+    """Input-output parameter setting for the backward direction."""
+
+    update: UpdateParameters = field(default_factory=UpdateParameters)
+    """Parameter for the update behavior."""
+
+    def as_bindings(self) -> devices.AnalogTileParameter:
+        """Return a representation of this instance as a simulator bindings object."""
+        return tile_parameters_to_bindings(self)
+
+
+@dataclass
+class UnitCellRPUConfig:
+    """Configuration for an analog (unit cell) resistive processing unit."""
+
+    bindings_class: ClassVar[Type] = devices.AnalogTileParameter
+
+    device: UnitCellDevice = field(default_factory=UnitCellDevice)
+    """Parameters that modify the behavior of the pulsed device."""
+
+    forward: IOParameters = field(default_factory=IOParameters)
+    """Input-output parameter setting for the forward direction."""
+
+    backward: BackwardIOParameters = field(
+        default_factory=BackwardIOParameters)
+    """Input-output parameter setting for the backward direction."""
+
+    update: UpdateParameters = field(default_factory=UpdateParameters)
+    """Parameter for the update behavior."""
+
+    def as_bindings(self) -> devices.AnalogTileParameter:
+        """Return a representation of this instance as a simulator bindings object."""
+        return tile_parameters_to_bindings(self)
